requirements.txt
docx2txt==0.8
fastapi==0.111.0
uvicorn==0.30.1
pymongo==4.9.1
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13
requests==2.32.3
spacy==3.7.6
pydantic==2.9.2
instructor==1.5.2
openai==1.51.2
python-dotenv==1.0.0